Use double for timing and memory metrics
Change observability timing and memory-related metrics from nullable long to nullable double to support fractional milliseconds and MB values. Updates include interfaces, NodeMetrics, ObservabilityCollector, execution observer, logging sinks (log format and LoggerMessage signatures), related docs and examples, and test adjustments to match the new types and precision. Logs now emit durations/processor times with sub-millisecond precision (F3) and sink scopes updated accordingly.

+++ src/NPipeline/Observability/IObservabilityCollector.cs
@@ -31,7 +31,7 @@ public interface IObservabilityCollector
     /// <param name="processorTimeMs">The processor time used in milliseconds.</param>
     /// <param name="pipelineName">The name of the pipeline this node belongs to. Null for top-level pipelines.</param>
     void RecordNodeEnd(string nodeId, DateTimeOffset timestamp, bool success, Guid pipelineId, Exception? exception = null, double? peakMemoryMb = null,
-        long? processorTimeMs = null, string? pipelineName = null);
+        double? processorTimeMs = null, string? pipelineName = null);
 
     /// <summary>
     ///     Records item processing metrics for a node.
