+++ tests/NPipeline.Extensions.Observability.Tests/ErrorScenarioTests.cs
@@ -517,7 +517,7 @@ public sealed class ErrorScenarioTests
         Exception? exception = null,
         int retryCount = 0,
         long? peakMemoryMb = null,
-        long? processorTimeMs = null,
+        double? processorTimeMs = null,
         double? throughputItemsPerSec = null,
         double? averageItemProcessingMs = null,
         long itemsProcessed = 100,
